+++ juego/caballo.py
@@ -5,8 +5,6 @@ class Caballo(Piece):
     white_str  = "♘"
     black_str  = "♞"
 
-    # def __str__(self):
-    #     if self.__color__ == "BLANCO":
-    #         return "♘"
-    #     else:
-    #         return "♞"
+    def valid_move(self, from_row, from_col, to_row, to_col):    
+        return (abs(from_row - to_row), abs(from_col - to_col)) in [(2, 1), (1, 2)]
+