+++ juego/reina.py
@@ -5,8 +5,20 @@ class Reina(Piece):
     white_str  = "♕"
     black_str  = "♛"
 
-    # def __str__(self):
-    #     if self.__color__ == "BLANCO":
-    #         return "♕"
-    #     else:
-    #         return "♛"
+    def valid_move(self, from_row, from_col, to_row, to_col):
+        # Movimiento como torre (horizontal o vertical)
+        if from_col == to_col or from_row == to_row: #DUDA
+            # Validación para movimiento horizontal o vertical
+            if not self._is_path_clear_straight(from_row, from_col, to_row, to_col):
+                return False
+            return True
+
+        # Movimiento como alfil (diagonal)
+        if abs(from_row - to_row) == abs(from_col - to_col):
+            # Validación para movimiento diagonal
+            if not self._is_path_clear_diagonal(from_row, from_col, to_row, to_col):
+                return False #DUDA
+            return True
+
+        # No es un movimiento válido para la reina
+        return False