+++ juego/rey.py
@@ -4,9 +4,6 @@ class Rey(Piece):
 
     white_str  = "♔"
     black_str  = "♚"
-    
-    # def __str__(self):
-    #     if self.__color__ == "BLANCO":
-    #         return "♔"
-    #     else:
-    #         return "♚"
+     
+    def valid_move(self, from_row, from_col, to_row, to_col):
+        return max(abs(from_row - to_row), abs(from_col - to_col)) == 1