+++ main.py
@@ -1,7 +1,6 @@
 
 from juego.chess import Chess
-class InvalidMove(Exception):
-    ...
+from juego.exceptions import InvalidMove
 
 def main():
     chess = Chess()
@@ -15,8 +14,8 @@ def play(chess):
        print("turno: ", chess.turn)
        from_row = int(input("From row:  "))
        from_col = int(input("From col:  "))
-       to_row = int(input("From row:  "))
-       to_col = int(input("From col:  "))
+       to_row = int(input("To row:  "))
+       to_col = int(input("To col:  "))
        chess.move(
             from_row,
             from_col,