+++ test/test_torre.py
@@ -48,7 +48,39 @@ class TestTorre(unittest.TestCase):
                 possibles,
                 [(4,3), (4,2), (4,1), (4,0)]
         )        
+    def test_move_vertical_desc_with_own_piece(self):
+        board = Board()
+        board.set_piece(6, 1, Peon("WHITE", board))
+        torre = Torre("WHITE", board)
+        board.set_piece(4, 1, Torre)
+        possibles = torre.possible_positions_vd(4, 1)
+        self.assertEqual(
+            possibles,
+            [(5, 1)]
+        )
 
+    def test_move_vertical_desc_with_not_own_piece(self):
+        board = Board()
+        board.set_piece(6, 1, Peon("BLACK", board))
+        torre = Torre("WHITE", board)
+        board.set_piece(4, 1, Torre)
+        possibles = torre.possible_positions_vd(4, 1)
+        self.assertEqual(
+            possibles,
+            [(5, 1), (6, 1)]
+        )
+
+    def test_move_diagonal_desc(self):
+        board = Board()
+        torre = board.get_piece(col=0, row=0)
+        is_possible = torre.valid_positions(
+            from_row=0,
+            from_col=0,
+            to_row=1,
+            to_col=1,
+        )
+
+        self.assertFalse(is_possible)